packages/pig-agent-tools/pyproject.toml
[project]
name = "pig-agent-tools"
version = "0.0.1"
description = "Tool extensions for pig-agent-core"
authors = [{ name = "Pig Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pig-agent-core>=0.0.4",
]

[project.optional-dependencies]
web = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "tavily-python>=0.3.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pig_agent_tools"]
